backend/requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<3.17
djangorestframework-simplejwt>=5.3,<5.6
django-cors-headers>=4.3,<4.7
django-filter>=24.2,<25.0
python-dotenv>=1.0,<2.0
dj-database-url>=2.1,<3.0

# Optional Postgres driver — installed only when DATABASE_URL points at Postgres.
# psycopg[binary]>=3.1,<4.0

# Tests
pytest>=8.2,<9.0
pytest-django>=4.8,<5.0
